Chapter3 no1~3 clear

+++ algorithm/chapter1/no7.java
@@ -6,11 +6,11 @@ public class no7 {
     public static void main(String[] args){
         
         Scanner scanint = new Scanner(System.in);
-        int a, b;
-        a = scanint.nextInt();
-        b = scanint.nextInt();
+        Double a, b;
+        a = scanint.nextDouble();
+        b = scanint.nextDouble();
 
         scanint.close();
-        System.out.println(a / b);
+        System.out.println(a/b);
     };
 };
